Retelling func

+++ front/app/lib/DataBases/RelationalDB.ts
@@ -1,17 +1,20 @@
 import {
-  SummaryRecordSchema,
-  SummarySchema,
-  TextMessageRecordSchema,
-  TextMessageRecordSchemaArray,
-  isSummarySchema,
-  isTextMessageRecordSchema,
-  isTextMessageRecordSchemaArray,
+  SummaryToRecordSchema,
+  SummaryReturnSchema,
+  TextMessageToRecordSchema,
+  TextMessageReturnRecordSchema,
+  isSummaryReturnSchemaArray,
+  isTextMessageReturnRecordSchemaArray,
 } from "@/app/utils/Validation";
 import { createClient } from "@supabase/supabase-js";
 import { z } from "zod/v4";
 
-if (!process.env.SUPABASE_URL || !process.env.SUPABASE_SERVICE_ROLE_KEY) {
-  throw new Error("Supabase config error");
+export function createSupabaseClient() {
+  if (!process.env.SUPABASE_URL) throw new Error();
+  return createClient<Database>(
+    process.env.SUPABASE_URL,
+    process.env.SUPABASE_SERVICE_ROLE_KEY!
+  );
 }
 
 export interface Database {
@@ -30,6 +33,7 @@ export interface Database {
           text: string;
           date: Date;
           messageType: string;
+          is_summarized: boolean;
         };
         Insert: {
           updateId: bigint;
@@ -42,6 +46,7 @@ export interface Database {
           text: string;
           date: Date;
           messageType: string;
+          is_summarized: boolean;
         };
         Update: {
           updateId?: bigint;
@@ -54,19 +59,48 @@ export interface Database {
           text?: string;
           date?: Date;
           messageType?: string;
+          is_summarized?: boolean;
+        };
+      };
+
+      summaries: {
+        Row: {
+          id: bigint;
+          chat_id: string;
+          participants: string;
+          text: string;
+          date_from: Date;
+          date_to: Date;
+          created_at: Date;
+          message_count: number;
+        };
+        Insert: {
+          chat_id: string;
+          participants: string;
+          text: string;
+          date_from: Date;
+          date_to: Date;
+          created_at: Date;
+          message_count: number;
+        };
+        Update: {
+          chat_id?: string;
+          participants?: string;
+          text?: string;
+          date_from?: Date;
+          date_to?: Date;
+          created_at?: Date;
+          message_count?: number;
         };
       };
     };
   };
 }
 
-export const supaClient = createClient<Database>(
-  process.env.SUPABASE_URL,
-  process.env.SUPABASE_SERVICE_ROLE_KEY
-);
+export const supaClient = createSupabaseClient();
 
 export async function saveMessagesRDB(
-  records: z.infer<typeof TextMessageRecordSchemaArray>
+  records: z.infer<typeof TextMessageToRecordSchema>[]
 ) {
   if (!records.length) {
     throw new Error("No records provided");
@@ -83,6 +117,7 @@ export async function saveMessagesRDB(
     text: record.text,
     date: record.date,
     messageType: record.messageType,
+    is_summarized: record.is_summarized,
   }));
 
   const { data, error } = await supaClient
@@ -94,7 +129,7 @@ export async function saveMessagesRDB(
     throw error;
   }
 
-  if (!isTextMessageRecordSchemaArray(data)) {
+  if (!isTextMessageReturnRecordSchemaArray(data)) {
     throw new Error("Supabase result doesn't match the schema");
   }
 
@@ -103,7 +138,7 @@ export async function saveMessagesRDB(
 
 export async function isHundredBatch(chatId: number): Promise<{
   result: boolean;
-  unsummarized?: z.infer<typeof TextMessageRecordSchemaArray>;
+  unsummarized?: z.infer<typeof TextMessageReturnRecordSchema>[];
 }> {
   const { data, error } = await supaClient
     .from("messages")
@@ -113,92 +148,118 @@ export async function isHundredBatch(chatId: number): Promise<{
 
   if (error) throw error;
 
-  if (!isTextMessageRecordSchemaArray(data)) {
+  if (!isTextMessageReturnRecordSchemaArray(data)) {
     throw new Error("Supabase result doesn't match the schema");
   }
 
   if (data.length < 100) {
-    console.log("Not enought messages to summarize. Done.");
+    console.log("✅ Not enought messages to summarize. Done.");
     return { result: false };
   }
   return { result: true, unsummarized: data };
 }
 
+// испралено
 export async function getUnsummarized(
   chatId: number
-): Promise<z.infer<typeof TextMessageRecordSchemaArray>> {
+): Promise<z.infer<typeof TextMessageReturnRecordSchema>[]> {
   const { data, error } = await supaClient
     .from("messages")
     .select("*")
     .eq("is_summarized", false)
-    .eq("chatId", String(chatId));
+    .eq("chatId", String(chatId))
+    .order("date", { ascending: true} );
 
   if (error) throw error;
 
-  if (!isTextMessageRecordSchemaArray(data)) {
+  if (!isTextMessageReturnRecordSchemaArray(data)) {
     throw new Error("Supabase result doesn't match the schema");
   }
 
   return data;
 }
 
+// исправлено
 export async function getLastSummary(
-  chat_id: number
-): Promise<z.infer<typeof SummarySchema>> {
+  chat_id: number,
+  count: number = 1
+): Promise<z.infer<typeof SummaryReturnSchema>[]> {
   const { data, error } = await supaClient
     .from("summaries")
     .select("*")
     .eq("chat_id", String(chat_id))
-    .order("created_at", { ascending: false})
-    .limit(1);
+    .order("created_at", { ascending: false })
+    .limit(count);
 
-  if(error) throw error;
+  if (error) throw error;
 
-  if(!isSummarySchema(data)) {
-    throw new Error("Supabase result doesn't match the schema")
+  console.log(data);
+  if (!isSummaryReturnSchemaArray(data)) {
+    throw new Error("Supabase result doesn't match the schema");
   }
 
   return data;
 }
 
+export async function getSummaryCount(chat_id: number): Promise<number> {
+  const { count, error } = await supaClient
+    .from("summaries")
+    .select("*", { count: "exact", head: true }) // <- ВАЖНО: head: true не возвращает сами данные
+    .eq("chat_id", String(chat_id));
+
+  if (error) throw error;
+
+  return count ?? 0;
+}
+
+
 export async function updateStatusSummarized(
-  summarized: z.infer<typeof TextMessageRecordSchemaArray>
+  summarized: z.infer<typeof TextMessageReturnRecordSchema>[]
 ) {
   const usedIds = summarized.map(
-    (msg: z.infer<typeof TextMessageRecordSchema>) => msg.messageId
+    (msg: z.infer<typeof TextMessageReturnRecordSchema>) => msg.id
   );
 
   await supaClient
     .from("messages")
     .update({ is_summarized: true })
-    .in("messageId", usedIds);
+    .in("id", usedIds);
 }
 
 export async function cleanRDB(chatId: number) {
   const { data: latest10, error: latestError } = await supaClient
     .from("messages")
     .select("*")
-    .eq("chatId", String(chatId))
+    .eq("chatId", chatId)
     .order("date", { ascending: false })
     .limit(10);
 
   if (latestError) throw latestError;
 
   const latest10Ids = latest10.map(
-    (m: z.infer<typeof TextMessageRecordSchema>) => m.messageId
+    (m: z.infer<typeof TextMessageReturnRecordSchema>) => m.id
   );
 
-  await supaClient
-    .from("message")
-    .delete()
-    .not("messageId", "in", latest10Ids)
-    .eq("chatId", chatId);
+  if (latest10Ids.length === 0) {
+    console.log("⚠️ No messages found to preserve. Skipping deletion.");
+    return;
+  }
 
-  console.log("Deleted old messages, kept latest 10.");
+  const { error: deleteError } = await supaClient
+    .from("messages")
+    .delete()
+    .not("id", "in", `(${latest10Ids.join(",")})`)
+    .eq("chatId", String(chatId));
+
+  if (deleteError) {
+    console.error("❌ Delete error:", deleteError);
+  }
+
+  console.log("☑️ Deleted old messages, kept latest 10.");
 }
 
 export async function saveSummaryRDB(
-  summaryPayload: z.infer<typeof SummaryRecordSchema>
+  summaryPayload: z.infer<typeof SummaryToRecordSchema>
 ): Promise<string> {
   const { data, error } = await supaClient
     .from("summaries")
@@ -209,11 +270,7 @@ export async function saveSummaryRDB(
 
   const summaryId: number = data[0].id;
 
-  console.log({
-    status: "ok",
-    message: "Summary inserted",
-    summaryId,
-  });
+  console.log("☑️ Save summary: summary inserted");
 
   return String(summaryId);
 }
@@ -221,7 +278,7 @@ export async function saveSummaryRDB(
 export async function summariesById(
   ids: string[],
   chatId: number
-): Promise<z.infer<typeof SummarySchema>[]> {
+): Promise<z.infer<typeof SummaryReturnSchema>[]> {
   const { data, error } = await supaClient
     .from("summaries")
     .select("*")
@@ -230,20 +287,20 @@ export async function summariesById(
 
   if (error) throw error;
 
-  console.log({
-    status: "ok",
-    message: "Summaries found",
-  });
+  console.log("☑️ Summaries by id: found");
+
+  if (!isSummaryReturnSchemaArray(data)) {
+    throw new Error("Data doesn't match SummaryReturnSchemaArray");
+  }
 
   return data;
 }
 
-// функцию, которая будет получать сырые последние сообщения
-
+// функция, которая будет получать сырые последние сообщения
 export async function getRawMessages(
   chatId: number,
   count: number
-): Promise<z.infer<typeof TextMessageRecordSchemaArray>> {
+): Promise<z.infer<typeof TextMessageReturnRecordSchema>[]> {
   const { data, error } = await supaClient
     .from("messages")
     .select("*")
@@ -255,7 +312,7 @@ export async function getRawMessages(
 
   if (!data) throw new Error("No data returned from Supabase.");
 
-  if (!isTextMessageRecordSchemaArray(data)) {
+  if (!isTextMessageReturnRecordSchemaArray(data)) {
     throw new Error("Data has incorrect type");
   }
